+++ src/app/api/models/strip.model.ts
@@ -1,0 +1,3 @@
+export interface StripData {
+    name: string;
+}