+++ src/app/components/image-configurator.tsx
@@ -1,11 +1,11 @@
-import { Portfolio } from "@/src/app/models/portfolio";
+import { ImageConfiguration, Portfolio } from "@/src/app/models/portfolio";
 import {
     Box,
     Button,
     Checkbox,
     CircularProgress,
     FormControl,
-    FormLabel,
+    FormControlLabel,
     InputLabel,
     MenuItem,
     Select,
@@ -14,34 +14,67 @@ import {
 import { useSnackbar } from "notistack";
 import * as React from "react";
 import { useEffect, useState } from "react";
+import { PaletteSelector } from "./palette-selector";
+import { StripSelector } from "./strip-selector";
 
 const customerPortfolioRoute: string = "/api/portfolio/my";
 const traderPortfolioRoute: string = "/api/portfolio";
 
 interface ImageConfiguratorProps {
     isTrader: boolean;
+    configurationId: string | null;
     imageId: string;
+    imageConfiguration: ImageConfiguration | null;
 }
 
-export function ImageConfigurator({ isTrader, imageId }: ImageConfiguratorProps) {
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
+const handleCreateConfiguration = async (data) => {
+    return await fetch(`/api/portfolio/configuration`, {
+        method: "POST",
+        body: JSON.stringify({
+            formData: data,
+        }),
+    });
+}
+
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-expect-error
+const handleSaveConfiguration = async (data) => {
+    return await fetch(`/api/portfolio/configuration`, {
+        method: "PUT",
+        body: JSON.stringify({
+            formData: data,
+        }),
+    });
+}
+
+export function ImageConfigurator({ isTrader, configurationId, imageId, imageConfiguration }: ImageConfiguratorProps) {
     const { enqueueSnackbar } = useSnackbar();
     const [loadingPortfolios, setLoadingPortfolios] = useState<boolean>(false);
     const [portfolios, setPortfolios] = useState<Portfolio[]>([]);
-    const [selectedPortfolio, setSelectedPortfolio] = useState<number | undefined>(undefined);
+    const [selectedPortfolio, setSelectedPortfolio] = useState<string | undefined>(undefined);
 
-    const [loadingPalettes, setLoadingPalettes] = useState<boolean>(false);
-    const [palettes, setPalettes] = useState<Palette[]>([]);
-    const [selectedPalette, setSelectedPalette] = useState("");
-
-    const [loadingStrips, setLoadingStrips] = useState<boolean>(false);
-    const [strips, setStrips] = useState<Strip[]>([]);
-    const [selectedStrip, setSelectedStrip] = useState("");
+    const [paletteId, setPaletteId] = useState<string | undefined>(undefined);
+    const [stripId, setStripId] = useState<string | undefined>(undefined);
+    const [passepartout, setPassepartout] = useState<boolean>(false);
 
     useEffect(() => {
         fetchPortfolios();
-        fetchPalettes();
-        fetchStrips();
     }, [])
+
+    useEffect(() => {
+        if (imageConfiguration) {
+            setSelectedPortfolio(imageConfiguration.portfolioId ?? undefined);
+            setPaletteId(imageConfiguration.paletteId ?? undefined);
+            setStripId(imageConfiguration.stripId ?? undefined);
+            setPassepartout(imageConfiguration.passepartout);
+        }
+    }, [imageConfiguration]);
+
+    const handleCheckPassepartout = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setPassepartout(event.target.checked);
+    };
 
     const fetchPortfolios = async () => {
         setLoadingPortfolios(true);
@@ -59,68 +92,35 @@ export function ImageConfigurator({ isTrader, imageId }: ImageConfiguratorProps)
         }
 
         setPortfolios(json["data"]);
-        if (json["data"].length > 0) {
+        if (json["data"].length > 0 && !selectedPortfolio) {
             setSelectedPortfolio(json["data"][0].id);
         }
     }
 
-    const fetchPalettes = async () => {
-        setLoadingPalettes(true);
-
-        const response = await fetch(`/api/palette`);
-
-        const json = await response.json();
-
-        setLoadingPalettes(false);
-
-        if (!response.ok) {
-            enqueueSnackbar(json.message, { variant: "error" });
-            setPalettes([]);
-            return;
-        }
-
-        setPalettes(json["data"]);
-        if (json["data"].length > 0) {
-            setSelectedPalette("-1");
-        }
-    }
-
-    const fetchStrips = async () => {
-        setLoadingStrips(true);
-
-        const response = await fetch(`/api/strip`);
-
-        const json = await response.json();
-
-        setLoadingStrips(false);
-
-        if (!response.ok) {
-            enqueueSnackbar(json.message, { variant: "error" });
-            setStrips([]);
-            return;
-        }
-
-        setStrips(json["data"]);
-        if (json["data"].length > 0) {
-            setSelectedStrip("-1");
-        }
-    }
-
     const handleSubmit = async (formData: FormData) => {
+        let portfolioId = selectedPortfolio;
+        if (imageConfiguration && imageConfiguration.portfolioId) {
+            portfolioId = imageConfiguration.portfolioId;
+        }
+
         const data = {
             imageId: imageId,
-            portfolioId: selectedPortfolio,
-            paletteId: selectedPalette === "-1" ? null : selectedPalette,
-            stripId: selectedStrip === "-1" ? null : selectedStrip,
+            portfolioId: portfolioId,
+            paletteId: !formData.get("palette-select") ? null : formData.get("palette-select"),
+            stripId: !formData.get("strip-select") ? null : formData.get("strip-select"),
             passepartout: formData.get("passepartout") === "on",
         }
 
-        const response = await fetch(`/api/portfolio/configuration`, {
-            method: "POST",
-            body: JSON.stringify({
-                formData: data,
-            }),
-        });
+        let response: Response;
+        if (configurationId) {
+            const putData = {
+                ...data,
+                id: configurationId,
+            }
+            response = await handleSaveConfiguration(putData);
+        } else {
+            response = await handleCreateConfiguration(data);
+        }
 
         const json = await response.json();
 
@@ -140,7 +140,7 @@ export function ImageConfigurator({ isTrader, imageId }: ImageConfiguratorProps)
         >
             {loadingPortfolios ? (<CircularProgress />) : portfolios && portfolios.length > 0 ? (
                 <>
-                    {isTrader && (
+                    {isTrader && !configurationId && (
                         <FormControl fullWidth>
                             <InputLabel id="portfolio-select">Mappe</InputLabel>
                             <Select
@@ -152,7 +152,7 @@ export function ImageConfigurator({ isTrader, imageId }: ImageConfiguratorProps)
                                 required
                                 value={selectedPortfolio}
                                 onChange={(event => {
-                                    setSelectedPortfolio(event.target.value as number);
+                                    setSelectedPortfolio(event.target.value as string);
                                 })}
                             >
                                 {portfolios.map((portfolio) => (
@@ -183,93 +183,55 @@ export function ImageConfigurator({ isTrader, imageId }: ImageConfiguratorProps)
 
             {portfolios.length > 0 && (
                 <>
-                    {loadingPalettes ? (<CircularProgress />) : (
-                        <FormControl fullWidth>
-                            <InputLabel id="palette-select">Palette</InputLabel>
-                            <Select
-                                label={"Palette"}
-                                id={"palette-select"}
-                                name={"palette-select"}
-                                autoFocus
-                                fullWidth
-                                required
-                                value={selectedPalette}
-                                onChange={(event => {
-                                    setSelectedPalette(event.target.value);
-                                })}
-                            >
-                                <MenuItem value="-1">
-                                    <em>Keine Auswahl</em>
-                                </MenuItem>
-
-                                {palettes.map((palette) => (
-                                    <MenuItem
-                                        key={palette.id}
-                                        value={palette.id}
-                                    >{palette.name}</MenuItem>
-                                ))}
-                            </Select>
-                        </FormControl>
-                    )}
-
-                    {loadingStrips ? (<CircularProgress />) : (
-                        <FormControl fullWidth>
-                            <InputLabel id="strip-color-select">Leistenfarbe</InputLabel>
-                            <Select
-                                label={"Leistenfarbe"}
-                                id={"strip-color-select"}
-                                name={"strip-color-select"}
-                                autoFocus
-                                fullWidth
-                                required
-                                value={selectedStrip}
-                                onChange={(event => {
-                                    setSelectedStrip(event.target.value);
-                                })}
-                            >
-                                <MenuItem value="-1">
-                                    <em>Keine Auswahl</em>
-                                </MenuItem>
-
-                                {strips.map((stripColor) => (
-                                    <MenuItem
-                                        key={stripColor.id}
-                                        value={stripColor.id}
-                                    >{stripColor.name}</MenuItem>
-                                ))}
-                            </Select>
-                        </FormControl>
-                    )}
-
-                    <FormControl fullWidth>
-                        <Box>
-                            <Checkbox
-                                id="passepartout"
-                                name="passepartout"
-                            />
-                            <FormLabel htmlFor="passepartout">Passepartout</FormLabel>
-                        </Box>
-                    </FormControl>
+                    <PaletteSelector
+                        onChange={(e) => setPaletteId(e.target.value)}
+                        selectedPalette={paletteId}
+                    />
+                    <StripSelector
+                        onChange={(e) => setStripId(e.target.value)}
+                        selectedStrip={stripId}
+                    />
+                    <FormControlLabel
+                        name={"passepartout"}
+                        control={<Checkbox
+                            checked={passepartout}
+                            onChange={handleCheckPassepartout}
+                        />}
+                        label="Passepartout"
+                    />
                 </>
             )}
 
-            <Box>
-                {portfolios.length > 0 ? (
-                    <Button
-                        variant={"contained"}
-                        type={"submit"}
-                    >
-                        Zu Mappe hinzufügen
-                    </Button>
-                ) : isTrader && (
-                    <Button
-                        variant={"contained"}
-                        href={"/portfolio/add"}
-                    >
-                        Neue Mappe erstellen
-                    </Button>
+            <>
+                {configurationId ? (
+                    <Box>
+                        <Button
+                            variant={"contained"}
+                            type={"submit"}
+                        >
+                            Speichern
+                        </Button>
+                    </Box>
+                ) : (
+                    <Box>
+                        {portfolios.length > 0 ? (
+                            <Button
+                                variant={"contained"}
+                                type={"submit"}
+                            >
+                                Zu Mappe hinzufügen
+                            </Button>
+                        ) : isTrader && (
+                            <Button
+                                variant={"contained"}
+                                href={"/portfolio/add"}
+                            >
+                                Neue Mappe erstellen
+                            </Button>
+                        )}
+                    </Box>
                 )}
-            </Box>
+            </>
         </form>
     )
 }