+++ src/app/components/palette-selector.tsx
@@ -1,0 +1,69 @@
+import { CircularProgress, FormControl, InputLabel, MenuItem, Select, SelectChangeEvent } from "@mui/material";
+import { enqueueSnackbar } from "notistack";
+import { useEffect, useState } from "react";
+
+interface PaletteSelectorProps {
+    required?: boolean;
+    onChange?: (event: SelectChangeEvent<string>) => void;
+    selectedPalette?: string | undefined;
+}
+
+export function PaletteSelector({
+    required = false,
+    onChange,
+    selectedPalette,
+}: PaletteSelectorProps) {
+    const [loading, setLoading] = useState<boolean>(true);
+    const [palettes, setPalettes] = useState<Palette[]>([]);
+
+    useEffect(() => {
+        (async () => {
+            setLoading(true);
+
+            const response = await fetch(`/api/palette`);
+
+            const json = await response.json();
+
+            setLoading(false);
+
+            if (!response.ok) {
+                enqueueSnackbar(json.message, { variant: "error" });
+                setPalettes([]);
+                return;
+            }
+
+            setPalettes(json["data"]);
+        })();
+    }, []);
+
+    return loading ? <CircularProgress /> : (
+        <FormControl fullWidth>
+            <InputLabel
+                id="palette-select"
+                shrink
+            >Palette</InputLabel>
+            <Select
+                label={"Palette"}
+                id={"palette-select"}
+                name={"palette-select"}
+                fullWidth
+                required={required}
+                displayEmpty
+                defaultValue={""}
+                value={selectedPalette}
+                onChange={onChange}
+            >
+                <MenuItem value={""}>
+                    <em>Keine Auswahl</em>
+                </MenuItem>
+
+                {palettes.map((palette) => (
+                    <MenuItem
+                        key={palette.id}
+                        value={palette.id}
+                    >{palette.name}</MenuItem>
+                ))}
+            </Select>
+        </FormControl>
+    );
+}