+++ src/app/models/portfolio.ts
@@ -4,3 +4,13 @@ export interface Portfolio {
     description: string | null;
     owner_id: string;
 }
+
+export interface ImageConfiguration {
+    id: string;
+    imageId: string;
+    portfolioId: string | null;
+    byTrader: boolean;
+    paletteId: string | null;
+    stripId: string | null;
+    passepartout: boolean;
+}